feat: 增加问卷发布功能

+++ src/components/QuestionCard/QuestionCard.tsx
@@ -1,22 +1,23 @@
 import './QuestionCard.css'
-import { FC, MouseEvent } from 'react'
+import { FC } from 'react'
 
 type PropsType = {
   id: string
   title: string
   isPublished: boolean
-  deleteQuestion: (id: string) => void
+  deleteQuestion?: (id: string) => void
+  publishQuestion?: (id: string) => void
 }
 
 const QuestionCard: FC<PropsType> = props => {
-  const { id, isPublished, title, deleteQuestion } = props
-  const hanldeListItemEdit = (e: MouseEvent<HTMLButtonElement>) => {
-    console.log('hanldeListItemEdit = ', e)
+  const { id, isPublished, title, deleteQuestion, publishQuestion } = props
+  const handleItemPublish = (id: string) => {
+    publishQuestion && publishQuestion(id)
   }
   const handleItemDelBtnClick = (id: string) => {
-    console.log('handleItemDelBtnClick = ')
-    deleteQuestion(id)
+    deleteQuestion && deleteQuestion(id)
   }
+
   return (
     <>
       <li className="list-item" key={id}>
@@ -25,7 +26,9 @@ const QuestionCard: FC<PropsType> = props => {
           {isPublished ? '已发布' : '未发布'}
         </span>
         <span className="btn-container">
-          <button onClick={e => hanldeListItemEdit(e)}>编辑问卷</button>
+          <button className="pub-btn" onClick={() => handleItemPublish(id)}>
+            {isPublished ? '下架问卷' : '发布问卷'}
+          </button>
           <button className="del-btn" onClick={() => handleItemDelBtnClick(id)}>
             删除问卷
           </button>
